--- src/autocomplete.css
@@ -1,41 +1,0 @@
-.search-container {
-   
-      display: flex;
-      align-items: center;
-      justify-content: center;
-    
- }
- 
- .search-container{
-    width: 100%;
-    position: relative;
-    display: flex;
-    flex-direction: column;
- }
- .search-container input{
-    width:40%;
-    height: 40px;
-    font-size: 25px;
- }
- 
- 
- .list-container li{
-   width:100%;
-    height: 40px;
-    list-style: none;
-    border: 1px solid gray;
-    font-size: 15px;
-    color:black;
-    background-color: white;
- }
-
-#list-container {
-   z-index:99999; 
-   position:absolute;
-    top:40px;
-    width:40%;
-
-}
-
-
-
