+++ src/components/currentloactionweather/weatherforyourlocation.css
@@ -13,7 +13,7 @@ body {
   background-repeat: no-repeat;
   box-shadow: 0px 7px 30px 2px rgba(0, 0, 0, 0.18);
 }
-img {
+svg {
   width: 60px;
   font-size: 17px;
   padding-right: 25px;
@@ -35,10 +35,12 @@ h1 {
     right: 10px;
     top: 10px; */
   text-align: right;
-  color: white;
 }
 .city {
   /* position: absolute; */
+}
+#time {
+  color: white !important;
 }
 
 .list-inline {
@@ -46,7 +48,7 @@ h1 {
   padding-right: 30px;
 }
 
-#tag {
+h6 {
   /* position: absolute;
   padding-top: 22px;
   margin:10px; */
@@ -54,32 +56,6 @@ h1 {
   position: absolute;
   bottom: 0;
   margin-bottom: 0;
-}
-
-.mb-5 {
-  position: absolute;
-  right: 0px;
-  top: 0px;
-  margin-top: 10px;
-  margin-right: 5px;
-}
-
-.Map {
-  color: white;
-  background-color: blue;
-  padding: 15px 10px;
-  display: inline-flex;
-  text-align: center;
-  align-items: center;
-  justify-content: center;
-  transform: (-50%, -50%);
-}
-#toggle-container {
-  width: 70px;
-  background-color: #c4c4c4;
-  cursor: pointer;
-  border-radius: 3px;
-  height: 32px;
 }
 
 .dialog-button {
@@ -104,13 +80,10 @@ h1 {
   background-color: orange;
   margin-left: 32px;
 }
-#f {
-  width: 400px;
-}
-
-@media only screen and (max-width: 768px) {
-  /* For mobile phones: */
-  #f {
-    width: 100%;
-  }
+#toggle-container {
+  width: 70px;
+  background-color: #c4c4c4;
+  cursor: pointer;
+  border-radius: 3px;
+  height: 32px;
 }
