+++ src/components/filtersoptions/hourly/hourly.css
@@ -38,9 +38,7 @@ body{
       text-align: right;
       
     }
-    .city{
-      /* position: absolute; */
-    }
+   
   #time {
     color:white !important;
   }
