+++ src/index.js
@@ -3,9 +3,10 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+  <App/>
   </React.StrictMode>,
   document.getElementById('root')
 );
