--- src/weatherforyourlocation.css
@@ -1,91 +1,0 @@
-body{
-  padding-top: 200px;
-}
-#card{
-  height: 100px;
-  margin-bottom: 80px;
-  width: 350px;
-  position: relative;
-  border-radius: 30px;
-  background-color:orange;
-  padding: 10px;
-  background-size: cover;
-  background-repeat: no-repeat;
-  box-shadow: 0px 7px 30px 2px rgba(0,0,0,0.18);
-}
- svg{
-      width: 60px;
-    font-size: 17px;
-    padding-right: 25px;
-  }
-  .x{
-    position: absolute;
-    bottom: 0;
-    right: 33px;
-  }
-
-
-  h1{
-    color:white;
-    position: absolute;
-    left: 10px;
-    bottom: 0px;
-  }
-  .date{
-    /* position: absolute;
-    right: 10px;
-    top: 10px; */
-    text-align: right;
-    
-  }
-  .city{
-    /* position: absolute; */
-  }
-#time {
-  color:white !important;
-}
-
-.list-inline{
-  margin-bottom: 0 !important;
-  padding-right: 30px;
-}
-
-h6{
-  /* position: absolute;
-  padding-top: 22px;
-  margin:10px; */
-  left: 33px;
-  position: absolute;
-  bottom:0;
-  margin-bottom:0;
-}
-
- .dialog-button {
-  font-size: 14px;
-    line-height: 16px;
-    font-weight: bold;
-    cursor: pointer;
-    background-color:orange;
-    color: white;
-    padding: 8px 8px;
-    border-radius: 18px;
-    min-width: 46px;
-    align-items: center;
-    width: 38px;
-    min-width: unset;
-    border-radius: 3px;
-    box-sizing: border-box;
-    transition: all 0.3s ease;
-  }
-  
-  .disabled {
-    background-color:orange;
-    margin-left: 32px;
-  }
-  #toggle-container {
-    width: 70px;
-    background-color: #c4c4c4;
-    cursor: pointer;
-    border-radius: 3px;
-    height: 32px;
-  }
